Move ut client from internal to pkg
